Use rake db:schema:load and db:seed to set up the db instead of db:migrate
So we all know db:migrate was never the answer all along. But I was considering whether we should use db:schema:loadand db:seed vs. loading an SQL file generated by db:structure:dump. The problem with db:structure:dump is that it includes a bunch of PostGIS tables, but not the complete PostGIS install. So the easy but not entirely clean answer is to go back to db:schema:load plus db:seed. Whenever we have execute statements in our migrations, we need to add those to seeds.rb or they won't end up in the database. Any views, triggers, plpgsql functions, or constraints need to go into this file in addition to the migration where they're created. This is not fun, but there's no other simple answer right now.